Refactor mg-pocos-de-caldas to use new BaseSiscamSpider
